+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added example for `terra-tabs` to test the focus traverse with interactive element within the tab content.
 
+* Changed
+  * Updated examples and tests for `terra-worklist-data-grid` to properly update row selection data.
+
 * Updated
   * Updated example for `terra-date-picker` with isInvalid prop.